fix a pile of issues with calling conventions on GLES2 platforms by directly calling the gl functions on these platforms, this required quite an overhaul in the headers and a lot of code gets disabled if USE_GLES2 to avoid referencing undefined constants and such...
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11266 d7cf8633-e32d-0410-b094-e92efae38249

+++ model_alias.c
@@ -3002,7 +3002,7 @@ void Mod_INTERQUAKEMODEL_Load(dp_model_t *mod, void *buffer, void *bufferend)
 	pend = (unsigned char *)bufferend;
 
 	if (pbase + sizeof(iqmheader_t) > pend)
-		Host_Error ("Mod_INTERQUAKEMODEL_Load: %s is not an Inter-Quake Model %d", loadmodel->name, pend - pbase);
+		Host_Error ("Mod_INTERQUAKEMODEL_Load: %s is not an Inter-Quake Model %d", loadmodel->name, (int)(pend - pbase));
 
 	// copy struct (otherwise it may be misaligned)
 	// LordHavoc: okay it's definitely not misaligned here, but for consistency...
